pyproject.toml
[project]
name = "agentforge"
version = "0.1.0"
description = "Transform job descriptions into deployable AI agent blueprints via PersonaNexus"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [{ name = "AgentForge Contributors" }]
keywords = ["ai", "agent", "agentforge", "job-description", "skills", "personanexus"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.0,<3.0",
    "pyyaml>=6.0.1,<7.0",
    "typer>=0.9,<1.0",
    "rich>=13.0,<14.0",
    "anthropic>=0.40,<1.0",
    "openai>=2.0,<3.0",
    "pymupdf>=1.24,<2.0",
    "python-docx>=1.0,<2.0",
    "personanexus>=1.4,<2.0",
]

[project.optional-dependencies]
web = [
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.30,<1.0",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0,<3.0",
    "alembic>=1.13,<2.0",
]

mcp = [
    "mcp>=1.0,<2.0",
]

langgraph = [
    "langgraph>=0.2,<1.0",
    "langchain-anthropic>=0.3,<1.0",
    "langchain-core>=0.3,<1.0",
]

[project.scripts]
agentforge = "agentforge.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentforge"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "--cov=agentforge --cov-report=term-missing"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "S", "C4", "SIM"]
ignore = [
    "S101",   # assert usage — acceptable in tests
    "S603",   # subprocess call — not applicable
    "S607",   # start process with partial path — not applicable
    "B904",   # raise in except without from — intentional for typer.Exit
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S110"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.8",
]
